+++ eva/src/components/dashBoardInside/dashTable.vue
@@ -7,7 +7,7 @@
         v-model="props.input"
         class="dash-table report-table"
         :headers="props.titles"
-        :items.sync="filteredTableData"
+        :items.sync="eventedTableData"
         :data-id="id"
         item-key="none"
         :hide-default-footer="props.hideFooter"
@@ -24,7 +24,7 @@
           v-slot:[`header.${title}`]="{ header }"
         >
           <v-menu :key="`${title + value}`" offset-y>
-            <template v-slot:activator="{ on, attrs }">
+            <template v-slot:activator>
               <v-menu z-index="100000" offset-y :close-on-content-click="false">
                 <template v-slot:activator="{ on, attrs }">
                   <v-icon
@@ -36,7 +36,7 @@
                     >{{ mdiMagnify }}</v-icon
                   >
                 </template>
-                <v-row v-if="value != 'string'">
+                <v-row v-if="value !== 'string'">
                   <v-col cols="6">
                     <v-select
                       :items="compare"
@@ -66,11 +66,35 @@
               </v-menu>
             </template>
           </v-menu>
-          <v-tooltip :key="header.value" bottom>
+          <v-tooltip :key="header.value + value" bottom>
             <template v-slot:activator="{ on }">
               <span v-on="on">{{ header.text }}</span>
             </template>
           </v-tooltip>
+        </template>
+        <template v-for="(title) in props.titles" v-slot:item="{ item }">
+          <tr
+            :key="title + item.rowIndex"
+            :style="item.rowColor && `background-color: ${item.rowColor}`"
+          >
+            <template v-for="(col, colIndex) in item">
+              <td
+                v-if="!excludeColumns.includes(colIndex)"
+                :key="colIndex"
+                class="text-start"
+                :style="
+                  (item.cellColor &&
+                    item.cellColor[colIndex] &&
+                    `background-color: ${item.cellColor[colIndex]}`) ||
+                  (item.columnColor &&
+                    item.columnColor[colIndex] &&
+                    `background-color: ${item.columnColor[colIndex]}`)
+                "
+              >
+                {{ col }}
+              </td>
+            </template>
+          </tr>
         </template>
       </v-data-table>
     </div>
@@ -81,7 +105,8 @@
 </template>
 
 <script>
-import { mdiMagnify } from '@mdi/js';
+import {mdiMagnify} from '@mdi/js';
+
 export default {
   props: {
     dataRestFrom: null,
@@ -119,69 +144,108 @@ export default {
       dataTitles: [],
       stringTitles: [],
       filters: {},
+      excludeColumns: ['rowColor', 'columnColor', 'cellColor', 'rowIndex'],
     };
   },
   computed: {
+    eventedTableData() {
+      const items = [...this.filteredTableData].map((item, index) => ({
+        ...item,
+        rowIndex: index,
+      }));
+      this.eventRows.forEach((event) => {
+        if (event.prop[0] === 'rowcolor') {
+          items.forEach((item) => {
+            if (this[event.compare](item[event.column], event.row)) {
+              item.rowColor = event.value[0];
+            }
+          });
+        }
+        if (event.prop[0] === 'columncolor') {
+          let isColumnMatch = items.reduce((acc, item) => {
+            return acc || this[event.compare](item[event.column], event.row);
+          }, false);
+          if (isColumnMatch) {
+            items.forEach((item) => {
+              if (!item.columnColor) {
+                item.columnColor = {};
+              }
+              item.columnColor[event.column] = event.value[0];
+            });
+          }
+        }
+        if (event.prop[0] === 'cellcolor') {
+          items.forEach((item) => {
+            if (this[event.compare](item[event.column], event.row)) {
+              if (!item.cellColor) {
+                item.cellColor = [];
+              }
+              item.cellColor[event.column] = event.value[0];
+            }
+          });
+        }
+      });
+      return items;
+    },
     filteredTableData() {
       let chooseSort = function (dataFormat, sortType, value) {
         if (dataFormat === 'date') {
           let sort;
           let parseDate = function (val) {
             let parts = val.split('.');
-            let date = new Date(
+            return new Date(
               Number(parts[0]),
               Number(parts[1]) - 1,
               Number(parts[2])
             );
-            return date;
           };
-          if (sortType == '>')
+          if (sortType === '>')
             sort = (el) => {
               let elDate = parseDate(el);
               let valueDate = parseDate(value);
               return valueDate < elDate;
             };
-          else if (sortType == '<')
+          else if (sortType === '<')
             sort = (el) => {
               let elDate = parseDate(el);
               let valueDate = parseDate(value);
               return valueDate > elDate;
             };
-          else if (sortType == '=')
+          else if (sortType === '=')
             sort = (el) => {
               let elDate = parseDate(el);
               let valueDate = parseDate(value);
-              return valueDate.getTime() == elDate.getTime();
+              return valueDate.getTime() === elDate.getTime();
             };
           return sort;
         } else if (dataFormat === 'number') {
           let sort;
-          if (sortType == '>')
+          if (sortType === '>')
             sort = (el) => {
               return +el > +value;
             };
-          else if (sortType == '<')
+          else if (sortType === '<')
             sort = (el) => {
               return +el < +value;
             };
-          else if (sortType == '=')
+          else if (sortType === '=')
             sort = (el) => {
-              return +value == +el;
+              return +value === +el;
             };
           return sort;
         } else if (dataFormat === 'string') {
           let sort;
-          if (sortType == '>')
+          if (sortType === '>')
             sort = (el) => {
               return el > value;
             };
-          else if (sortType == '<')
+          else if (sortType === '<')
             sort = (el) => {
               return el < value;
             };
-          else if (sortType == '=')
+          else if (sortType === '=')
             sort = (el) => {
-              return value == el;
+              return value === el;
             };
           return sort;
         }
@@ -197,35 +261,33 @@ export default {
         }
       }
       return temp;
-    },
 
+    },
     events() {
-      let events = this.$store.getters.getEvents({
+      return this.$store.getters.getEvents({
         idDash: this.idDash,
         event: 'OnDataCompare',
         element: this.id,
       });
-      return events;
     },
-    id: function () {
+    id() {
       return this.idFrom;
     },
-    idDash: function () {
+    idDash() {
       return this.idDashFrom;
     },
-    change: function () {
+    change() {
       if (!this.dataRestFrom || this.dataRestFrom.length === 0) {
-        this.props.itemsForTable = [];
-        this.props.nodata = true;
+        this.setNoData()
       } else {
         this.getDataAsynchrony(this.dataRestFrom);
       }
       return true;
     },
-    theme: function () {
+    theme() {
       return this.$store.getters.getTheme;
     },
-    height: function () {
+    height() {
       let otstup = 100;
       if (screen.width <= 1600) {
         otstup = 80;
@@ -239,10 +301,10 @@ export default {
       if (this.dataReport) {
         otstup = otstup - 30;
       }
-      let height = this.heightFrom - otstup; // 120 это размер блока с пагинацией таблицы + шапка с настройками самого блока
-      return height;
+       // 120 это размер блока с пагинацией таблицы + шапка с настройками самого блока
+      return this.heightFrom - otstup;
     },
-    lastResult: function () {
+    lastResult() {
       let options = this.$store.getters.getOptions({
         idDash: this.idDash,
         id: this.id,
@@ -253,9 +315,9 @@ export default {
   watch: {
     options: {
       deep: true,
-      handler(newValue, oldValue) {
+      handler(newValue) {
         if (newValue) this.createTitles(newValue.titles);
-      }
+      },
     },
     dataRestFrom: {
       deep: true,
@@ -279,6 +341,10 @@ export default {
     this.setEventColor();
   },
   methods: {
+    setNoData() {
+      this.props.itemsForTable = [];
+      this.props.nodata = true;
+    },
     indexTitles(oldVal) {
       let type = 'no';
       for (let [key, val] of Object.entries(oldVal[0])) {
@@ -298,7 +364,6 @@ export default {
     getType(title) {
       return this.typedTitles[title];
     },
-
     setFilterData(title, event, compare) {
       if (!this.filters[title]) this.filters[title] = {};
       if (compare === 'compare') {
@@ -312,8 +377,8 @@ export default {
       function isNumber(n) {
         return /^-?[\d.]+(?:e-?\d+)?$/.test(n);
       }
-      if (isNumber(val)) return true;
-      return false;
+      return isNumber(val);
+
     },
     checkForString(val) {
       return Object.prototype.toString.call(val) === '[object String]';
@@ -323,11 +388,11 @@ export default {
       let parts = val.split('.');
       if (parts.length < 3) return false;
       let result;
-      let mydate = new Date(parts[2], parts[1] - 1, parts[0]);
+      let mydate = new Date(+parts[2], +parts[1] - 1, +parts[0]);
       if (
-        parts[2] == mydate.getYear() &&
-        parts[1] - 1 == mydate.getMonth() &&
-        parts[0] == mydate.getDate()
+        +parts[2] === mydate.getYear() &&
+        +parts[1] - 1 === mydate.getMonth() &&
+        +parts[0] === mydate.getDate()
       ) {
         result = 0;
       } else {
@@ -335,7 +400,7 @@ export default {
       }
       return result;
     },
-    getDataAsynchrony: function (data) {
+    getDataAsynchrony(data) {
       let prom = new Promise((resolve) => {
         if (data.error) {
           this.props.message = data.error;
@@ -359,24 +424,29 @@ export default {
         this.props.itemsForTable = data;
       });
     },
-    createTitles: function (result) {
+    createTitles(result) {
       if (this.options?.titles) {
         let allTitles = Object.keys(this.dataRestFrom[0]);
-        this.props.titles = allTitles.map(x => ({
+        this.props.titles = allTitles.map((x) => ({
           text: x,
           value: x,
           sortable: true,
-          align: (this.options.titles.length === 0 || this.options.titles.includes(x)) ? undefined : ' d-none'
+          align:
+            this.options.titles.length === 0 || this.options.titles.includes(x)
+              ? undefined
+              : ' d-none',
         }));
       } else {
         if (result && result.length) {
           this.props.titles = Object.keys(result[0]).map((item) => {
-            return { text: item, value: item, sortable: true };
+            if (!this.excludeColumns.includes(item)) {
+              return { text: item, value: item, sortable: true };
+            }
           });
         }
       }
     },
-    createTockens: function (result) {
+    createTockens(result) {
       let captures = Object.keys(result[0]);
       this.props.actions.forEach((item, i) => {
         this.$set(this.props.actions[i], 'capture', captures);
@@ -387,145 +457,33 @@ export default {
         id: this.id,
       });
     },
-    setEventColor: function () {
-      if (this.eventRows.length > 0) {
-        for (let x of this.eventRows) {
-          x.style.background = '';
-          x.classList.remove('event');
-        }
-      }
-      this.eventRows = [];
-      let events = this.$store.getters.getEvents({
+    equals(a, b) {
+      return a === b;
+    },
+    over(a, b) {
+      return +a > +b;
+    },
+    less(a, b) {
+      return +a < +b;
+    },
+    in(a, b) {
+      return b.includes(a);
+    },
+    between(a, b) {
+      return b[0] < a || a < b[1];
+    },
+    setEventColor() {
+      this.eventRows = this.$store.getters.getEvents({
         idDash: this.idDash,
         event: 'OnDataCompare',
         element: this.id,
       });
-      let table, column;
-      let eventObj = {};
-      events.forEach((item, index) => {
-        eventObj[index] = {};
-        eventObj[index]['compare'] = item.compare;
-        eventObj[index]['column'] = item.column;
-        eventObj[index]['row'] = item.row;
-        eventObj[index]['color'] = item.value[0];
-        eventObj[index]['prop'] = item.prop[0];
-        table = this.$refs.table.$el;
-
-        if (
-          eventObj[index]['prop'] == 'rowcolor' ||
-          eventObj[index]['prop'] == 'columncolor' ||
-          eventObj[index]['prop'] == 'cellcolor'
-        ) {
-          let readyTh = setTimeout(
-            function tick() {
-              table.querySelectorAll('thead th').style =
-                'background-color: red';
-              if (table.querySelectorAll('thead th').length != 0) {
-                clearTimeout(readyTh);
-                let sp = 0;
-                table.querySelectorAll('thead span').forEach((itemSpan) => {
-                  if (itemSpan.innerText != 0) {
-                    if (itemSpan.innerText == eventObj[index]['column']) {
-                      column = sp;
-                    }
-                    sp++;
-                  }
-                });
-
-                table.querySelectorAll('tbody tr').forEach((itemRow) => {
-                  itemRow.querySelectorAll('td').forEach((itemTd, i) => {
-                    if (i == column) {
-                      // itemRow.style.backgroundColor = 'yellow';
-                      let needItem = null,
-                        row,
-                        k = -1;
-
-                      switch (eventObj[index]['compare']) {
-                        case 'equals':
-                          if (itemTd.innerText == eventObj[index]['row']) {
-                            needItem = itemRow;
-                          }
-
-                          break;
-                        case 'over':
-                          if (itemTd.innerText > eventObj[index]['row']) {
-                            needItem = itemRow;
-                          }
-                          break;
-                        case 'less':
-                          if (itemTd.innerText < eventObj[index]['row']) {
-                            needItem = itemRow;
-                          }
-                          break;
-                        case 'in':
-                          row = eventObj[index]['row']
-                            .replace(/\[|\]/g, '')
-                            .split(',');
-                          k = -1;
-                          row.forEach((rowValue) => {
-                            if (itemTd.innerText == rowValue) {
-                              k = 0;
-                            }
-                          });
-                          if (k != -1) {
-                            needItem = itemRow;
-                          }
-                          break;
-                        case 'between':
-                          row = eventObj[index]['row']
-                            .replace(/\[|\]/g, '')
-                            .split(',');
-                          if (
-                            itemTd.innerText > row[0] &&
-                            itemTd.innerText < row[1]
-                          ) {
-                            needItem = itemRow;
-                          }
-                          break;
-                      }
-                      if (needItem != null && this.eventRows) {
-                        needItem.classList.add('event');
-                        this.eventRows.push(needItem);
-                      }
-                    }
-                  });
-                });
-
-                if (table.querySelectorAll('.event').length > 0) {
-                  if (item.prop[0] == 'rowcolor') {
-                    let rows = table.querySelectorAll('.event');
-                    rows.forEach((res) => {
-                      res.style.background = eventObj[index]['color'];
-                      //res.style.color = this.color.back;
-                    });
-                  } else if (item.prop[0] == 'cellcolor') {
-                    table.querySelectorAll('.event').forEach((res) => {
-                      res.children[column].style.background =
-                        eventObj[index]['color'];
-                      //res.children[column].style.color = this.color.back;
-                    });
-                  } else if (item.prop[0] == 'columncolor') {
-                    table.querySelectorAll('tbody tr').forEach((itemRow) => {
-                      itemRow.children[column].style.background =
-                        eventObj[index]['color'];
-                      //itemRow.children[column].style.color = this.color.back;
-                    });
-                  }
-                }
-              } else {
-                readyTh = setTimeout(tick, 100);
-              }
-            }.bind(this),
-            0
-          );
-        }
-      });
     },
-    selectRow: function () {
+    selectRow() {
       document
         .querySelector(`[data-id=${this.id}]`)
         .addEventListener('click', (event) => {
-          if (event.target.tagName.toLowerCase() == 'td') {
+          if (event.target.tagName.toLowerCase() === 'td') {
             if (event.target.parentElement.classList.contains('selected')) {
               event.target.parentElement.classList.remove('selected');
             } else {
@@ -544,14 +502,14 @@ export default {
 
             let cellRowIndex = Array.from(
               event.target.parentElement.childNodes
-            ).findIndex((item) => item == event.target);
+            ).findIndex((item) => item === event.target);
 
             let tokens = this.$store.getters.getTockens(this.idDash);
 
             tokens.forEach((token) => {
               if (
-                token.elem == this.id &&
-                token.action == 'click' &&
+                token.elem === this.id &&
+                token.action === 'click' &&
                 headers[cellRowIndex] === token.capture
               ) {
                 let value = event.target.textContent;
@@ -571,14 +529,14 @@ export default {
               partelement: 'row',
             });
 
-            if (events.length != 0) {
+            if (events.length !== 0) {
               events.forEach((item) => {
-                if (item.action == 'set') {
+                if (item.action === 'set') {
                   this.$store.commit('letEventSet', {
                     events: events,
                     idDash: this.idDash,
                   });
-                } else if (item.action == 'go') {
+                } else if (item.action === 'go') {
                   this.$store.commit('letEventGo', {
                     event: item,
                     idDash: this.idDash,