+++ eva/src/store/store.js
@@ -15,7 +15,7 @@ export default {
     },
   },
   actions: {
-    async actionGetElementSelected({ commit, state, getters }, element) {
+    async actionGetElementSelected({ commit, getters }, element) {
       const selected = getters.getElementSelected({
         idDash: element.idDash,
         id: element.id,
@@ -33,70 +33,76 @@ export default {
     // проверяет и создает объект в хранилище для настроек
     // path - это idDash либо произвольное место хранения настроек например research
     // element - например multiLine, multiLine-2, table, table-2 ...
-    async prepareSettingsStore({ commit, state, getters }, { path, element }) {
-      const idExists = (element !== undefined && element !== '')
-      if (typeof path === 'number'){
-        path = path.toString()
+    async prepareSettingsStore({ state }, { path, element }) {
+      const idExists = element !== undefined && element !== '';
+      if (typeof path === 'number') {
+        path = path.toString();
       }
       if (!state[path.toString()]) {
-        state[path] = {}
+        state[path] = {};
         if (idExists) {
-          state[path][element] = {}
+          state[path][element] = {};
         }
       }
       if (idExists) {
         if (!state[path][element].modalSettings) {
           state[path][element].modalSettings = {
             element: '',
-            status: false
-          }
+            status: false,
+          };
         }
         if (!state[path][element].options) {
-          state[path][element].options = {}
+          state[path][element].options = {};
         }
       } else {
         if (!state[path].modalSettings) {
           state[path].modalSettings = {
             element: '',
-            status: false
-          }
+            status: false,
+          };
         }
         if (!state[path].options) {
-          state[path].options = {}
+          state[path].options = {};
         }
       }
     },
 
     // сохранение настроек
-    async saveSettingsToPath({ commit, state, getters, dispatch }, { path, element, options }) {
-      await dispatch('prepareSettingsStore', { path, element })
-      commit('setOptions', { idDash: path, id: element, options })
+    async saveSettingsToPath(
+      { commit, getters, dispatch },
+      { path, element, options }
+    ) {
+      await dispatch('prepareSettingsStore', { path, element });
+      commit('setOptions', { idDash: path, id: element, options });
       return await getters.getOptions({ idDash: path, id: element });
     },
 
     // получение настроек
-    async getSettingsByPath({ commit, state, getters, dispatch }, { path, element }) {
-      await dispatch('prepareSettingsStore', { path, element })
+    async getSettingsByPath({ getters, dispatch }, { path, element }) {
+      await dispatch('prepareSettingsStore', { path, element });
       return await getters.getOptions({ idDash: path, id: element });
     },
 
     // открыть окно настроек
     // произвольный вызов this.$store.dispatch("openModalSettings", { path: 'research', element: 'multiLine' });
-    async openModalSettings({ commit, state, getters, dispatch }, { path, element, titles }) {
-      await dispatch('prepareSettingsStore', { path, element })
+    async openModalSettings({ commit, dispatch }, { path, element, titles }) {
+      await dispatch('prepareSettingsStore', { path, element });
       return await commit('setModalSettings', {
         idDash: path,
         element,
         status: true,
-        titles
+        titles,
       });
     },
 
     // закрыть окно настроек
-    async closeModalSettings({ commit, state, getters, dispatch }, { path }) {
-      return await commit('setModalSettings',  { idDash: path, status: false, id: '' } );
-    }
-
+    async closeModalSettings({ commit }, { path }) {
+      return await commit('setModalSettings', {
+        idDash: path,
+        status: false,
+        id: '',
+      });
+    },
   },
   mutations: {
     setNameDash: (state, newName) => {
@@ -278,12 +284,12 @@ export default {
           let data = null;
           eventAll.forEach((item) => {
             // пробегаемся по всем событиям
-
+            let value, k;
             switch (
               state[idDash].events[item].compare // проверяем какое именно событие должно произойти
             ) {
               case 'equals':
-                let value = state[idDash].tockens[id].value.replace(/\s/g, ''); // в случаях когда нужно сравнить значения токена по равенству,
+                value = state[idDash].tockens[id].value.replace(/\s/g, ''); // в случаях когда нужно сравнить значения токена по равенству,
                 // это может быть строка, а значит нужно обрезать пробелы, чтобы сравнение было корректным
                 if (value === state[idDash].events[item].tokenval) {
                   // сравниваем значения в событии и значение токена
@@ -320,7 +326,7 @@ export default {
                 data = state[idDash].events[item].tokenval
                   .replace(/\[|\]/g, '')
                   .split(','); // отбрасываем скобки массива и разбиваем на масив элемнетов по запятой
-                let k = -1;
+                k = -1;
                 data.forEach((item) => {
                   // каждое значнеие нужно сравнить со значением в событии
                   if (
@@ -733,7 +739,7 @@ export default {
         }
       });
       if (options.titles) {
-        state[options.idDash][options.id].selectedTableTitles = options.titles; // deprecated
+        state[options.idDash][options.id].selectedTableTitles = options.titles;
       }
     },
 
@@ -843,7 +849,7 @@ export default {
         if (typeof itemValue === 'string' && itemValue.indexOf('$') !== -1) {
           itemValue = itemValue.replace(/\$/g, '');
 
-          tockens.forEach((tockenDeep, l) => {
+          tockens.forEach((tockenDeep) => {
             if (tockenDeep.name == itemValue) {
               values[k] = tockenDeep.value;
             }
@@ -890,10 +896,10 @@ export default {
       const currentTab = event.event.tab || state[id]?.currentTab;
       const isTabMode = state[id]?.tabs;
       let lastEl;
-      const isGoToTabExsits = state[id]?.tabList.find((el) => {
-        lastEl = el;
-        return el.id == event.event.tab;
-      });
+      // const isGoToTabExsits = state[id]?.tabList.find((el) => {
+      //   lastEl = el;
+      //   return el.id == event.event.tab;
+      // });
       if (!options?.openNewScreen) {
         if (!isTabMode) {
           event.route.push(`/dashboards/${id}/1`);
@@ -940,11 +946,11 @@ export default {
                   item.sid,
                   id
                 );
-                responseDB.then((result) => {
-                  let refresh = event.store.getters.refreshElements(
-                    id,
-                    item.sid
-                  );
+                responseDB.then(() => {
+                  // let refresh = event.store.getters.refreshElements(
+                  //   id,
+                  //   item.sid
+                  // );
                   event.store.commit('setLoading', {
                     search: item.sid,
                     idDash: id,
@@ -1161,7 +1167,7 @@ export default {
       for (let part of filter.parts) {
         state[filter.idDash].stashedFilterParts.push({
           ...part,
-          values: [...part.values],
+          values: part.values ? [...part.values] : [],
         });
       }
     },
@@ -1584,7 +1590,7 @@ export default {
     putIntoDB() {
       // затем полученные данные нужно положить в indexed db
       return (result, sid, idDash) => {
-        return new Promise((resolve, reject) => {
+        return new Promise((resolve) => {
           let db = null;
           let id = idDash;
           let key = `${id}-${sid}`;
@@ -1603,7 +1609,7 @@ export default {
               db.createObjectStore('searches'); // create it
             }
 
-            request.onsuccess = (event) => {
+            request.onsuccess = () => {
               db = request.result;
               console.log('success: ' + db);
 
@@ -1611,7 +1617,7 @@ export default {
             };
           };
 
-          request.onsuccess = (event) => {
+          request.onsuccess = () => {
             db = request.result;
 
             setTransaction(db, result, key, idDash);
@@ -1667,11 +1673,11 @@ export default {
 
         let request = indexedDB.open('EVA', 1);
 
-        request.onerror = function (event) {
+        request.onerror = function () {
           console.log('error: ');
         };
 
-        request.onsuccess = (event) => {
+        request.onsuccess = () => {
           db = request.result;
 
           let transaction = db.transaction('searches', 'readwrite'); // (1)
@@ -1785,7 +1791,7 @@ export default {
       // получаем скриншот страницы
       return (id) => {
         if (!id.id) {
-          return []
+          return [];
         }
         if (!state[id.idDash][id.id].options) {
           Vue.set(state[id.idDash][id.id], 'options', {});
@@ -1865,7 +1871,7 @@ export default {
     getTheme(state) {
       return state.theme.settings;
     },
-    getThemeBack(state) {
+    getThemeBack() {
       return () => {
         return rest.getThemeBack(restAuth);
       };
@@ -1891,7 +1897,7 @@ export default {
     },
     saveDashboard: () => {
       return (dash) => {
-        return new Promise((resolve, reject) => {
+        return new Promise((resolve) => {
           let response = restAuth.setEssence({
             formData: JSON.stringify(dash),
             essence: 'dash',
@@ -1971,7 +1977,7 @@ export default {
     },
     checkDataSearch: () => {
       return (sid) => {
-        return new Promise((resolve, reject) => {
+        return new Promise((resolve) => {
           let db = null;
 
           let request = indexedDB.open('EVA', 1);
@@ -1988,7 +1994,7 @@ export default {
               db.createObjectStore('searches'); // create it
             }
 
-            request.onsuccess = (event) => {
+            request.onsuccess = () => {
               db = request.result;
               // this.alreadyDB = request.result;
               console.log('success: ' + db);
@@ -1997,7 +2003,7 @@ export default {
             };
           };
 
-          request.onsuccess = (event) => {
+          request.onsuccess = () => {
             db = request.result;
 
             setTransaction(db);
@@ -2011,7 +2017,7 @@ export default {
 
             let query = searches.get(sid); // (3) return store.get('Ire Aderinokun');
 
-            query.onsuccess = (event) => {
+            query.onsuccess = () => {
               // (4)
 
               if (query.result) {